test: changes on the end to end tests
